requirements.txt
streamlit==1.24.0
opencv-contrib-python==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
pyyaml==6.0
python-dotenv==1.0.0